requirement.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
CacheControl==0.14.4
celery==5.3.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
filelock==3.24.3
firebase-admin==6.3.0
fonttools==4.61.1
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
httplib2==0.31.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
kombu==5.6.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
msgpack==1.1.2
networkx==3.4.2
numpy==1.26.4
opencv-python==4.8.1.78
packaging==26.0
pandas==2.3.3
Pillow==10.1.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
rsa==4.9.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
thop==0.1.1.post2209072238
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.0.196
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
